+++ lib/src/cli_app.dart
@@ -179,7 +179,7 @@ class CliApp {
     _argsMap["particleExamples"] = "Install Particle Example";
 
     //Demos
-    _argsMap["moppiFlower"] = "Install Flower Demo";
+    _argsMap["moppiFlowerExamples"] = "Install Flower Demo";
 
     _argsMap.forEach((flag, description) {
       argParser.addFlag(flag, negatable: false, help: description);